Fix: libclang search order (system first), size_t type for Linux, pytest testpaths

+++ src/transpilers/frontends/cpp/parser/preprocess.py
@@ -112,7 +112,7 @@ namespace std {
     template <class T> class deque {};
     template <typename R, typename... A> class function {};
     class exception {};
-    using size_t = unsigned long long;
+    using size_t = unsigned long;
     using ptrdiff_t = long;
     class string { public:
         string(); string(const char*); string(const string&);
@@ -206,8 +206,8 @@ namespace std {
     template <typename T, typename U> T* static_cast_helper(U*);
     template <typename T, typename U> T* dynamic_cast_helper(U*);
 }
-void* operator new(unsigned long long, void*);
-void* operator new(unsigned long long);
+void* operator new(unsigned long, void*);
+void* operator new(unsigned long);
 void operator delete(void*);
 
 // TU-scope `using` declarations so user code that writes bare
